requirements.txt
pygame==2.1.2
grpcio==1.68.0
grpcio-tools==1.68.0
protobuf>=5.26.1,<6.0dev